BCB-1577 moved validate_sgcn_input.py to pysgcn so it will be picked up with the pysgcn package on the pipeline deploy
