+++ pysgcn/validate_sgcn_input.py
